refacto: move updateLobbyInContext
